refactor(canvas): split shape types into files

Copilot-Session: 16f9a514-5f9e-4563-bde4-3bf48e654807
